ragg-package {ragg}	R Documentation

ragg: Graphic Devices Based on AGG

Description

Anti-Grain Geometry (AGG) is a high-quality and high-performance 2D drawing library. The 'ragg' package provides a set of graphic devices based on AGG to use as alternative to the raster devices provided through the 'grDevices' package.

Author(s)

Maintainer: Thomas Lin Pedersen thomas.pedersen@posit.co (ORCID)

Authors:

Maxim Shemanarev (Author of AGG) [copyright holder]

Other contributors:

Tony Juricic tonygeek@yahoo.com (Contributor to AGG) [contributor, copyright holder]

Milan Marusinec milan@marusinec.sk (Contributor to AGG) [contributor, copyright holder]

Spencer Garrett (Contributor to AGG) [contributor]

Posit, PBC [copyright holder, funder]

See Also

Useful links:

https://ragg.r-lib.org

https://github.com/r-lib/ragg

Report bugs at https://github.com/r-lib/ragg/issues

[Package ragg version 1.2.6 Index]